Modify the running topo of script test_fdb_mac_move.py. (#7923)
Description of PR
In PR #7789, I add mark pytest.mark.topology('any') before the script test_fdb_mac_move.py, thus this script will run on all of topologies. But in this script, we need to get the VLAN information of topo. On no-vlan topo, such as t1, it will raise an error KeyError: 'VLAN'. So in this PR, modify the mark and this script will only run on T0.

What is the motivation for this PR?
In PR #7789, I add mark pytest.mark.topology('any') before the script test_fdb_mac_move.py, thus this script will run on all of topologies. But in this script, we need to get the VLAN information of topo. On no-vlan topo, such as t1, it will raise an error KeyError: 'VLAN'. So in this PR, modify the mark and this script will only run on T0.

How did you do it?
Modify the pytest mark before the script.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>