Update configlet with the missing port_channel config information used in add rack script #7872
Config used to add rack via port channel there is missing the elements "tpid" and "lacp_key". These elements are generated as part of the original config but not part of the add_rack config.

What is the motivation for this PR?
This PR is closing the gap in the config used to add T0 rack with T1 rack via port channel

How did you verify/test it?
Added the missing config in configlet and re-ran the test script to verify the difference between the original config_db and the config_db post add rack process

Signed-off-by: yenlu <yenlu@cisco.com>