Add library and sample scripts for running ansible in pure python (#7792)
What is the motivation for this PR?
This change experimented the idea of running ansible in pure python. With the classes and functions defined in ansible_hosts.py and devices.py, we can use pure python to run any ansible module on any hosts. The required inputs are ansible inventory file and optional variable files. We can view it as a python version of playbook. Comparing with ansible playbook, we can take advantage of a real programming language. The drawback is that python programming experience is required.

Comparing with pytest-ansible, we do not need pytest. This design supports some ansible features not supported by pytest-ansible:
* fork: Pytest-ansible does not support running ansible modules in parallel. This design uses ansible's builtin forking capability to run modules in parallel.
* module attributes: Ansible supports additional module attributes that can be specified for each task in playbook. These module attributes can affect execution of the modules, for example "become", "async", etc. Pytest-ansible does not support these attributes. With this design, we can use keyword argument module_attrs to specify module attributes while calling an ansible module.

Not all ansible's builtin features are supported by this design. For example:
* Notify and event handler. (We can use python's libs to support that)

This idea is still new. I haven't figured out all of its potentials and limitations. Feedbacks, suggestions and contributions are more than welcome!

How did you do it?
* Added basic class AnsibleHosts and AnsibleHost
* Added class SonicHosts
* Added helper functions like init_xxx_hosts
* Added script ansible/upgrade_sonic.py as replacement of ansible/upgrade_sonic.yml
* Added script .azure-pipelines/upgrade_image.py which can be called in nightly test.

How did you verify/test it?
Test run the upgrade_sonic.py and upgrade_image.py scripts.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>