Merged PR 2349: Do not start service systemd-resolved in pretest
Do not start service systemd-resolved in pretest

The pretest case test_connect_to_internal_nameserver adds
internal DNS server to /etc/resolv.conf and restarts service
systemd-resolved. The systemd-resolved service is disabled
by default. We are not suppose to start it. Otherwise, the
/etc/resolv.conf file may be overwritten.

This change removed the code for restarting systemd-resolved.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>