Fixes test error after Python3 migration in test_vnet_vxlan.py (#7971)
Summary:
Fixes test error after Python3 migration in test_vnet_vxlan.py

How:
The issue is due to the use of minigraph_vlans.keys()[0] and minigraph_vlans[topo_vlan].members[1:] in the Jinja2 template, which are not supported in Python 3.
In Python 3, dict.keys() returns a dictionary view object, which does not support indexing. Instead, you can convert it to a list and then access the element using indexing.

Signed-off-by: yenlu <yenlu@cisco.com>