Switch to use Jinja's join() function for adding commas as needed
In Jinja 2.9, the behavior of `set` was fixed such that using `set`
within scopes (such as those started by for-loops) would not affect
variables set outside of that scope. Prior to this version, this could
be used to update variables outside of that scope. This means that
adding commas to separate multiple items based on this method would
fail.

Switch to the `join()` method defined by Jinja, which is meant exactly
for this purpose.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>