fix type mismatch error (#7952)
### Approach
#### What is the motivation for this PR?
To fix the TypeError raised in dualtor/test_tor_ecn.py.
packet.show() method returns Unicode strings, which cannot be written to a BytesIO object in Python3. Therefore, the TypeError is raised. 

#### How did you do it?
To fix this issue, we need to use StringIO instead of BytesIO to capture the output of the packet.show() method as strings.

#### How did you verify/test it?
Run the test locally.
Signed-off-by: zitingguo-ms <zitingguo@microsoft.com>