Change SONiC neighbor topo config to use exabgp and only config_db (#7894)
* Change SONiC neighbor topo config to use exabgp and only config_db

For T0 topology with SONiC neighbor, the T1 peers don't connect to
exabgp on the PTF server and use a mix of config_db and manual FRR
configuration. (For T2, it does look like it connects to exabgp.) This
results in common SONiC commands such as `show ip bgp summary` not
working because of the manual FRR configuration, along with the config
not being saved in a single file. In addition, because of the lack of
connection to exabgp, prefixes are not advertised, and many test cases
that involve IO likely can't be tested.

Fix this by specifying all of the configuration within config_db, along
with adding configuration to connect to exabgp. Also, remove the manual
frr/zebra configuration files.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Fix commas

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Remove trailing whitespace

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Make the BGP peer config generation more robust

Don't assume that the first BGP peer specified is IPv4, and that the
second one is IPv6. In addtion, normalize addresses to be lowercase so
that neighbor listing in `show ip bgp summary` works.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Change IPv6 detection method

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Change comma handling for interfaces and IPv4/IPv6 addresses

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Switch to use Jinja's join() function for adding commas as needed

In Jinja 2.9, the behavior of `set` was fixed such that using `set`
within scopes (such as those started by for-loops) would not affect
variables set outside of that scope. Prior to this version, this could
be used to update variables outside of that scope. This means that
adding commas to separate multiple items based on this method would
fail.

Switch to the `join()` method defined by Jinja, which is meant exactly
for this purpose.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>