[configlet] Fix configlet add_rack test (#7942)
Summary: Adjust configlet add_rack test by changing the format of PORTCHANNEL and QUEUE.

What is the motivation for this PR?
The add_rack test always fail on some platforms.
The issue was caused by a recent minigraph change. The @member will no longer be kept in PORTCHANNEL table.
Also, make the change on QUEUE range to 8 which was limited to 7.

How did you do it?
Change the format of portchannel and QUEUE.

How did you verify/test it?
Test it on DUT which failed the test.