[Mellanox] Skip ttl expired drop counters test on new Nvidia platforms (#7999)
Nvidia(Mellanox) platforms doesn't support the ttl expired drop counters, the test case has been skipped by the keyword "x86_64-mlnx_msn" in platform name. But for the new Nvidia platforms, the name will be changed to "x86_64-nvidia_snXXXX", so need to add this keyword in the skip condition.