[pre-commit] Fix style issues in scripts or files under `.hooks`, `docs` folder (#8016)
What is the motivation for this PR?
pre-commit is a static analysis tool introduced recently. This tool is
not able to do diff-only check. It checks the whole files touched by PR.
We can't blame PR author for legacy issues. That's why currently the
pre-commit check is only optional.

To ensure that we can make pre-commit a mandatory check for PRs
submitted to this repository, we need to fix all the legacy issues
complained by pre-commit.

How did you do it?
This change fixes the style issues of scripts or files under `.hooks`,
`docs` folder

Signed-off-by: yenlu <yenlu@cisco.com>