Change topology mark for test_snmp_default_route and test_snmp_loopback (#7620)
What is the motivation for this PR?
How did you do it?
-- test_snmp_default_route.py
The original topology mark is 'any', but there is no default route in the 'ptf' topology.
-- test_snmp_loopback.py
Use "nbrhosts" module to get the VM information.
But it will cause the error if the topology does not have VM(ex: ptf32).

How did you verify/test it?
Can skip two test cases on ptf32 topology.

Signed-off-by: yenlu <yenlu@cisco.com>