Ignore two syslog messages when doing reboot/config_reload of T2 chassis with DNX chipset (#6056)
What is the motivation for this PR?
When we reboot / config_reload on T2 chassis cards, we see 2 error messages in the linecards

During config_reload/reboot of linecard, LAGS are deleted, but ports are up,
and we get mac learning events from SAI to orchagent which is in middle of cleanup and doesn't have the right data.
This causes error message like Failed to get port by bridge port ID

reboot/config_reload on supoervisor will cause all the fabric links in the linecard to
bounce which results in SAI sending messages orchagent regarding the fabric port state change.
However, in linecards in T2 chassis, there is modelling of fabric ports in orchagent. Thus, orchagent generates
error message indication to port object found for the port.

How did you do it?
Created ignore_t2_syslog_msgs method in dut_utils.py
Called the method in cases when we do rebbot or reload
- tests/common/config_reload.py
- tests/common/reboot.py
- tests/platform_tests/api/test_module.py

How did you verify/test it?
Tested the reboot/reload testcases