Merged PR 2422: Remove docker registry from "no_proxy"
Remove docker registry from "no_proxy"

The legacy docker registry is in a VM that can be directly
accessed from starlab. That's why we added it to the "no_proxy"
list in env.yml. The new registry is running in Azure. Proxy
is required for accessing it.

This change removed the registry URL from the "no_proxy" list.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>