Set sonic-mgmt docker hostname (#8058)
What is the motivation for this PR?
Make the sonic-mgmt container hostname align with the container name on the host.

How did you do it?

How did you verify/test it?
Run the setup-container.sh script and verify the created container hostname is the same as the -n argument given to the script.

Signed-off-by: Lawrence Lee <lawlee@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>