Fix for Dockerfile smell DL3059 (#8044)
The Dockerfile placed at "spytest/containers/keysight-ubuntu18/Dockerfile" contains the best practice violation DL3059 detected by the hadolint tool.

The smell DL3059 occurs if there are multiple consecutive RUN instructions. Each RUN will correspond to a layer of the final Docker image, thus is recommended to compact them to reduce the number of layers for the final image.
This pull request proposes a fix for that smell generated by my fixing tool. The patch was manually verified before opening the pull request. To fix this smell, specifically, The consecutive RUN instructions have been chained until a comment or a different instruction appears.

This change is only aimed at fixing that specific smell. If the fix is not valid or useful, please briefly indicate the reason and suggestions for possible improvements.

Signed-off-by: Giovanni Rosa <g.rosa1@studenti.unimol.it>
Signed-off-by: yenlu <yenlu@cisco.com>