Fix the code for stopping exabgp in cleanup.sh.j2
After the exabgp processes are grouped to exabgpv4
and exabgpv6 in ptf container, the existing code for
stopping exabgp processes in cleanup.sh.j2 was broken.

The exabgp processes cannot be stopped before stopping
the PTF containers. Then, there is a chance that the server
got stuck if we run the cleanup.sh script.

This change updated the cleanup.sh.j2 template to stop
the exabgp processes by group names.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>