[Python3 migration]Fix test failure in ecmp
Fixes [ecmp/inner_hashing] | TypeError: a bytes-like object is required, not 'str' #7902

How did you do it?
When creating NamedTemporaryFile object, add mode argument and set to 'w+', which indicates that the file should be opened in text mode for reading and writing. The write method is then called with the string.

How did you verify/test it?
Manually run the test cases and didn't see the error.

Signed-off-by: yenlu <yenlu@cisco.com>