Merged PR 2450: Adjust dx010 t1 pipeline to run 202012
Why I did this?
We decide to qualify dx010 t1 by 202012 other than 202205 because known issue of platform, which are addressed in 202012 already.

How I did this?
To disable the pipeline of dx010 t1 for other platforms, and add them to 202012 by adding the yml file.

How to verify it?
N/A