Revert bad cherry-pick "Fix test_everflow_dscp_with_policer for mellanox vendor (#8050)" (#8120)
What is the motivation for this PR?
This reverts commit d6707b1.
PR #8050 was not cleanly cherry-picked to 202012 branch and introduced KeyError: 'everflow_dut'.

How did you do it?
This PR is to revert it firslty. I'll try cherry-pick again. If cherry-pick is not working, may need to ask the original author to help create a separate PR to 202012 branch .

How did you verify/test it?