Bug fix to get iface_namingmode tests to work on multi-asic linecards of a T2 chassis (#6528)
What is the motivation for this PR?
Currently, the following test isfailing on multi-asic linecards on a T2 VoQ chassis.

show pfc counters: This fails as the Inband and Recycle ports show in the 'show pfc counters' table while they are not expected.
How did you do it?
show pfc counters:
Treat Inband and Recycle ports on a VoQ chassis as internal ports. Since the inband and recycle ports do not show up in 'show' commands unless we specify '-d all' option, they should be treated similar to internal ports.
How did you verify/test it?
Ran the test against a multi-asic linecard in a T2 VoQ chassis.

Signed-off-by: yenlu <yenlu@cisco.com>