[advanced-reboot] Fix issue that prevented device shutdown and full r… (#8108)
The test_advanced_reboot fails on master branch when it fails to reboot the device, and fails w/ incorrect error: DUT hasn't shutdown in 300 seconds.
In reality the device is not rebooted at all as the test library crashes while doing so. The error is apparent when traceback is enabled in the code.

This PR fixes the error (introduced after change in #7817), added traceback for better debug support.
Also, this fixes a part where reboot timing data is not printed entirely as the logs path can't be parsed due to a change in
#7049

Tested on physical testbed

Signed-off-by: yenlu <yenlu@cisco.com>