Add conditional mark sort check to pre-commit hook
What is the motivation for this PR?
In conditional mark, testcases should be sorted from A to Z, and there
is no support to check it in PR test.

How did you do it?
Add conditional mark sort check to pre-commit hook, if testcases are not
sorted, pre-commit will fail, need to fix it before PR get merged.