Add conditional mark sort check to pre-commit hook (#8162)
What is the motivation for this PR?
In conditional mark, testcases should be sorted from A to Z, and there is no support to check it in PR test.

How did you do it?
Add a pre-commit hook under sonic-mgmt repo to check conditional mark sort, if testcases are not sorted, pre-commit will fail, need to fix it before PR get merged.

Signed-off-by: yenlu <yenlu@cisco.com>