Fix hard coded PSUs in conn_graph_facts (#8215)
What is the motivation for this PR?
The conn_graph_facts.py module hard coded PSUs of a devices as:
    ['PSU1', 'PSU2', 'PSU3', 'PSU4']
This assumes that a device can have at most 4 PSUs. For chassis devices, this is not true. A chassis device may have more than 4 PSUs.

How did you do it?
This change fixed the issue by dynamically get configured PSUs from connection graph.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>