Update dscp remapping cases for Nvidia platforms
This commit contains the fix for two test cases in DSCP remapping test. The reason that we need this fix is Nvidia implementation for DSCP remapping is different from Community due to the limitation that Nvidia platforms can't do the remap directly on tunnel.
For test case test_tunnel_decap_dscp_to_pg_mapping:
1. Add a new mapping file tunnel_qos_map_nvidia.json and use it when testing Nvidia platforms.
2. Use cell size 144 for Nvidia platforms.
3. Use outer DSCP instead of inner DSCP.
4. Add some logs for easier debugging.
For the xoff test case:
1. Add test data in qos.yml file for 4600C.
2. Stop packet aging in buffer before test.
3. Skip filling the leak on Nvidia platforms for it is not needed.
4. Use cell size 144 and packet length 300, small packets will cause descriptor exhaustion.