Improve stability for test test_reload_config.py::test_reload_configuration_checks (#8139)
The test occasionally fails with error

This error occurs in the check check_interfaces_config_service_status when running command "systemctl status interfaces-config.service" on the dut right after a reboot.
The root cause is after a reboot, when the system is not fully loaded, sometimes the command "systemctl status interfaces-config.service" will return with a rc 3. And in this case, the ansible command module fails, and then dump_ansible_results(self.results)).encode().decode("utf-8") is called in the errors.py. But in the output there is some Unicode characters that cannot be encoded as ascii.

We can just ignore the ansible module error because the output is checked against a regexp in check_interfaces_config_service_status, and there is a wait_until for this check.