Fix test_auto_negotiation issue (#8041)
For nvidia device, when buffer model is dynamic, the headroom size is related to the speed and cable length. When speed is bigger or equal to 400G and cable length is bigger, due to the ASIC headroom limitation, there will be some log errors.

So we need to change the cable length to one smaller value such as 50m.

Signed-off-by: yenlu <yenlu@cisco.com>