[dualtor-aa] Fix `test_fdb[ethernet]` (#8158)
Approach
What is the motivation for this PR?
Fix test_fdb[ethernet] failure on dualtor-aa testbed.

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
For pkt_type etherent, the packets to populate the fdb are ECMPed to
either ToRs, neither of the ToRs could have the full set of fdb entries.
And as the I/O also gets ECMPed in the validation step, the I/O could
not be guaranteed to be forwarded to the ToR that has the corresponding
fdb entry.

Let's fall back to active-standby mode to test this scenario for
active-active ports.

How did you verify/test it?
Run on dualtor-aa testbed.

Signed-off-by: yenlu <yenlu@cisco.com>