Merged PR 2623: Synchronize lock/release testbed action to TestbedV2
**Why**
The TBShare API will be deprecated and replaced by testbedV2, currently synchronizing the lock/release state of both for a while, then migrate the API to testbedV2 only.

**How**
Only for the current lock/release operation successfully, sync this operation to TestbedV2, and the sync operation will not block the current lock/release action.