Clearnup reference to github.com/Azure for SONiC repositories (#8305)
What is the motivation for this PR?
All the SONiC repositories have been migrated to github.com/sonic-net. We need to correct the links referring to old URLs.

How did you do it?
This change is to cleanup a few references to github.com/Azure.

How did you verify/test it?
Document change. No test required.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>