platform_tests/api: Fix iptables rule removal after reboot test (#8163)
After a reboot test, the iptables rule inserted via the
start_platform_api_server fixture will no longer exist and removing it will
cause an error. The code does currently check if the server is running but
that is not sufficient because the platform API server will come back up
after restart. Thus, just ignore the error so that cleanup can continue and
the test will pass, given that the iptables rule is expected to sometimes
no longer be present.