[TestbedV2]Add a script to get duts version (#8334)
Description of PR
In TestbedV2, we need a script to get duts version for futher process. So, we add this script using new sonichost, and this script will return a list contains the output of show version command.

What is the motivation for this PR?
In TestbedV2, we need a script to get duts version for futher process. So, we add this script using new sonichost, and this script will return a list contains the output of show version command.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>