Update iface_namingmode/test_iface_namingmode.py::TestShowQueue::test_show_queue_counters to work correctly on VOQ chassis (#7922)
This is a fix for issue
https://github.com/sonic-net/sonic-mgmt/issues/7875

The test is adapted to account for the BUFFER_QUEUE table format
differences in VOQ chassis compared to other architecture and the
existence of system ports in the BUFFER_QUEUE table from
other cards in the chassis.

How did you do it?
Adapted the BUFFER_QUEUE key parsing logic to always use
the last 2 fields to extract the interface name instead of assuming
the first 2 fields.

Ignore non-local system ports from other cards in the validation
logic.

How did you verify/test it?
Verifed that the the test passes on VOQ chassis and another fixed XGS broadcom SKU.

Signed-off-by: yenlu <yenlu@cisco.com>