Fixing single asic for GB. (#8326)
Adding a check for single-asic for cisco-8000 T2 platforms.

For cisco-8000 platforms (currently named as "gb" in qos.yaml), we have to fallback to the qos.yaml parameters listed for topo-any in case of single-ASIC tests(src & dst ASICS are same). For multi-asic and multi-DUT tests, the qos.yaml provides the parameters under topo-t2.

So we add a logic to check if the current test is single-asic and use topo-any for the parameters, and if not, use topo-t2.