Merged PR 2541: Create container qualification pipeline
Sonic k8s will need a pipeline for container qualification. We will leverage functions of the image qualification pipeline, the container qualification will add a more step to upgrade a specific container after image upgrade, and then run the same test cases as image qualification. This pipeline will be run on demand when there is a new feature container to be qualified.