Fixed issue on slow CPUs(mostly on DPUs) when we have failure due to 2 techsupport processes running but only one expected (#8312)
The problem happens only on slow CPU devices(observed on DPU devices with ARM CPU). Auto-techsupport tries to trigger a new techsupport process - which checks if the existing process is running and if the process already running it exits immediately, but in slow CPU devices - it exits not immediately, and as result we may see 2 techsupport processes running(second one will exit in few sec).
Fixed test logic - now we check that we have 1 or more processes running.
If we have more processes running - at the end of the test we will do validation for new techsupport files(we still expect only one new techsupport dump file).

Signed-off-by: Petro Pikh <petrop@nvidia.com>
Signed-off-by: yenlu <yenlu@cisco.com>