Avoid error if collect SSH cipher cases when DUT is offline
The SSH cipher tests are parameterized based on supported
SSH ciphers on DUT. In test item collection phase, scripts
need to run commands on DUT to get the supported ciphers.
If we just want to run pytest with "--collect-only", the collection
will fail if DUT is offline.

This change is to enhance parameterization of the SSH cipher tests.
Before connect to DUT, firstly check if the `--collect-only` pytest
argument is specified. If yes, do not try to connect to DUT. Just
return default supported ciphers list. Then pytest collection will
not raise error if DUT is offline.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>