[vSONiC]: Support peer device shwere a port channel isn't used
Some topologies (such as t2_2lc_min_ports-masic) don't use port channels
at all on the neighbor devices, but the current configdb template
assumes that there is a Port-Channel1 interface. This is for the
purposes of determining what the local address is for a BGP session.

Change this to getting all of the local IP addresses defined, and
determining which subnet the peer address falls into. There are cases
where this could break (one interface has an IP address and subnet
defined that is within a less-specific IP address and subnet for a
different interface), but this should hopefully cover all of our cases.

Fixes #8109.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>