Avoid using http proxy during announce-routes
The announce-routes step is to populate the exabgp processes running
in PTF container with predefined routes by sending HTTP POST request
from sonic-mgmt container to PTF container. The python requests tool
is used for sending HTTP requests. In case http_proxy or https_proxy
environment variables are set, the requests tool will use the proxy
to send the request. However, the proxy is not needed in this case
because usually sonic-mgmt container usually has direct access to PTF
container.

This change disables the proxy by explicitly setting proxies to None
while calling the `requests.post` method.

Another change is to use socket instead of "curl" command to wait for
the http port of exabgp processes running in PTF container to be ready.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>