Fix VLAN failure test case tc4 and tc5 (#8157)
What is the motivation for this PR?
The timeout of few sec (5 seconds here) for the expected packet is right way, before declaring failure.

The below two VLAN test case failed.
It is observed that the expected packet were not recieved as first packet.
Analysis the packet capture, the expected packet were recieved at
packet number 7, instead of packet number 0.

    **vlan/test_vlan.py::test_vlan_tc4_tagged_unicast
    vlan/test_vlan.py::test_vlan_tc5_untagged_unicast**

What is the motivation for this PR?
Add timeout for VLAN test case.
Timeout was not correctly done in the call to dp_poll via the verify_packet_any_fixed()

How did you do it?
Capture the packets on PTF to verify expected packets are received within the timeout value.

How did you verify/test it?
Run test with the fix to verify expected packets is received.

Any platform specific information?
This should be applicable to all platform, and not just for Cisco platform.