Avoid using http proxy during announce-routes (#8420)
What is the motivation for this PR?
The announce-routes step is to populate the exabgp processes running in PTF container with predefined routes by sending HTTP POST request from sonic-mgmt container to PTF container. The python requests tool is used for sending HTTP requests. In case http_proxy or https_proxy environment variables are set, the requests tool will use the proxy to send the request. However, the proxy is not needed in this case because usually sonic-mgmt container has direct access to PTF container.

How did you do it?
This change disables the proxy by explicitly setting proxies to None while calling the requests.post method.

Another change is to use socket instead of "curl" command to wait for the http port of exabgp processes running in PTF container to be ready.

How did you verify/test it?
Tested testbed-cli.sh announce-routes with and without http_proxy environment variable.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>