[202205][dualtor][active-active] Enable decap tests
Approach
What is the motivation for this PR?
Add decap test support on dualtor-aa topology.

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
As for active-active dualtor, the upstream packet from ptf is ECMPed to both ToRs, so we cannot tell which ToR could receive the IPinIP packet.
So let's use Loopback2 to bind decap rules for dualtor-aa topology, so IPinIP packets from ptf with dest IP as Loopback2 address will be duplicated to both ToRs, and could be decapsulated by both ToRs.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>