Add static route to send traffic for multi-asic and multi-dut tests for chassis-packets (#8365)
These changes are needed to make the multi-asic test variants to work over chassis-packet platforms

chassis-packet platform  uses internal iBGP routing to determine forwarding path for packets destined to other ASICs(same linecard or different linecards) via the Fabric Card. Since the qos-sai setup removes BGP, this routing information is lost. We need to setup a static route in the source-asic so that the packets can be forwarded to the destination ASIC.