Add PR test job to collect all pytest cases (#8390)
What is the motivation for this PR?
Traditionally, if we run pretest using mark option like -m pretest, then all test scripts in the repository are collected to find out which tests have mark pretest. If any of the script in repository has importing issue, the issue would be exposed while collecting the test.
Currently, the PR testing does not run the pretest by mark. It runs the test_pretest.py script as an ordinary script. Then if a script has importing issue, the issue cannot be detected.

How did you do it?
This change added a PR test job to run pytest command with --collect-only option. Then pytest will try to collect all the test only without actually executing any of them.
In case any script has import issue, this collect only check is able to detect it.