add support to alternative timestamps format (#8458)
This PR adds support to alternative timestamp format on platform test conftest script, such as 2023-05-30T09:29:04.578444+00:00
Which occurs when upgrading from 202012 to 202211.
This fix is compatible only with python 3.2+ as it uses datetime.timezone() type
This test is included in the python3 test list, so no compatibility issues are expected.

What is the motivation for this PR?
Fixing ValueError received due to lack of support to the mentioned timestamp, which occurs during the upgrade from 202012 to 202211

How did you do it?
Added a new format constant, and fixed the try-except flow, so it would try this format as the last resort.

How did you verify/test it?
ran the test on python3.

Signed-off-by: yenlu <yenlu@cisco.com>