Fix caching issue status in conditional mark
The conditional mark plugin supports querying Github for issues
status to determine adding a mark or not.
When there is network issue, querying Github for issue status
will fail after connection timeout. PR #6313 added caching
for issue status. However, the caching seems not really working
because multiprocessing is used for querying issue status. The
returned result is a proxy to Manager object, not really a dict
expected by the caching code.

This fix casted the issue querying result to dict before returning
it back to the caching code.

This change also added timeout value for sending requests
to Github querying issue status.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>