Increase the tolerance of noise packets in dualtor tunnel monintor (#8486)
The noise packets are considered in the dualtor tunnel traffic monitor, but the tolerance is too small sometimes.
Currently the tolerance in queue_stats_check function is 10% of the total sent packets.
If only a few packets are sent, the tolerance will be too small. For example, the default packets number is 10, in this case only 1 noise packet is allowed.
Sometimes dualtor tests fails due to noise packets.
So, increase the noise packets tolerance in queue check to at least 10 packets.

What is the motivation for this PR?
Fix the statistical failure in the queue check of dualtor tunnel traffic monitor.

How did you do it?
Increase the noise packets tolerance in queue check to at least 10 packets.

How did you verify/test it?
By automation, all related tests passed.

Signed-off-by: yenlu <yenlu@cisco.com>