Add logging to the image upgrade module
The customized ansible module reduce_and_add_sonic_images is for
upgrading image of SONiC device. When this module failed, it is
usually hard to debug.

This change updated this module to use the `config_module_logging`
function in module debug utilities. With this change, this module
would be able to generate log to `/tmp` folder on SONiC device for
easier debugging.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>