Improve devutils to specify output file (#8503)
What is the motivation for this PR?
The devutils tool supports output content in json. To save the json format output content to a file, we can run the command and redirect the output content to a file. However, the libraries called by this tool may generate some error messages to stdout. Those error messages may be redirected to the json file too. Then the saved json file could be invalid and cause subsequent parsing issue.

How did you do it?
This change enhanced the devutils script to support specifying an output file. Only valid content will be saved to the output file. Error messages will only be sent to console.

When the new argument is not specified, behavior of the devutils tool remains the same.

How did you verify/test it?
Test run the script with and without -o argument.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>