Enhance  testQosSaiQWatermarkAllPorts and  testQosSaiSharedReservationSize  for T2 (#8377)
The following fixes are present in this commit:

Adjust the qos.yaml numbers for WRR to be allowed by the limitations in ptf. The current number of packets is too large for ptf queue to process reliably.
Move some prelimilary checks for testQosSaiSharedReservationSize to a fixture. The fixture will force the test to skip if there are not enough ports to run the test, instead of failing as it does now.
Changed a recently added testcase(testQosSaiQWatermarkAllPorts) to become mutliasic aware.
As @abdosi requested, changed the brute-force function to use TCP instead of UDP. Also changed all the calling code.
Also some code cleanup, no logical changes - like reformatting some areas, and pring useful messages instead of asserting blankly.

Co-authored-by: abdosi <58047199+abdosi@users.noreply.github.com>