Fix the label for cherry-pick in workflow yaml
The workflow added for auto cherry-pick expects label like
`Approved for $branch branch`.
For the sonic-mgmt repository, people are used to use label like
`Request for $branch branch`.

Because of this difference, the cherry-pick actions are not really working.

This fix updated the expected labels in workflow yaml to `Request for $branch branch`.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>